Assignment1/requirements.txt
glob2==0.7
keras==2.12.0
matplotlib==3.5.2
numpy==1.22.4
PyYAML==6.0
PyYAML==6.0
scikit_image==0.19.2
scikit_learn==1.1.1
tensorflow==2.12.0
